+++ api.js
@@ -50,7 +50,7 @@ async function getProduct(req, res, next) {
  * @param {object} req 
  * @param {object} res 
  */
-async function createProduct(req, res) {
+async function createProduct (req, res, ) {
   const product = await Products.create(req.body)
   res.json(product)
 }
@@ -61,39 +61,31 @@ async function editProduct (req, res, next) {
   res.json(product)
 }
 
-/**
- * Delete a product
- * @param {Request} req
- * @param {Response} res
- * @param {NextFunction} next
- */
+
 async function deleteProduct (req, res, next) {
   const response = await Products.destroy(req.params.id)
   res.json(response)
 }
+
 async function createOrder (req, res, next) {
-const order = await Orders.create(req.body)
-res.json(orders)
+  const order = await Orders.create(req.body)
+  res.json(order)
 }
 
-/**
-* List orders
-* @param {Request} req
-* @param {Response} res
-* @param {NextFunction} next
-*/
+
 async function listOrders (req, res, next) {
-const { offset = 0, limit = 25, productId, status } = req.query
+  const { offset = 0, limit = 25, productId, status } = req.query
 
-const orders = await Orders.list({ 
-  offset: Number(offset), 
-  limit: Number(limit),
-  productId, 
-  status 
-})
+  const orders = await Orders.list({ 
+    offset: Number(offset), 
+    limit: Number(limit),
+    productId, 
+    status 
+  })
 
-res.json(orders)
+  res.json(orders)
 }
+
 module.exports = autoCatch({
   handleRoot,
   listProducts,
