+++ app.js
@@ -19,9 +19,10 @@ app.get('/products/:id', api.getProduct)
 app.put('/products/:id', api.editProduct)
 app.delete('/products/:id', api.deleteProduct)
 app.post('/products', api.createProduct)
-// Boot the server
+
 
 app.get('/orders', api.listOrders)
-app.get('/orders/', api.createOrder)
+app.post('/orders', api.createOrder)
+// Boot the server
 app.listen(port, () => console.log(`Server listening on port ${port}`))
